Updated Navigation component and app styles
Modified the Navigation.svelte file to improve the look and feel of the navigation bar by adding a backdrop filter, adjusting the height, and changing the margin of unordered list items. Additionally, updated the app.pcss file to change the background image from a .png to a .jpg file for enhanced loading speed and performance.